# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Authentication API",
    "version": "1.0.0",
    "description": "API for user authentication and data seeding"
  },
  "paths": {
    "/signup": {
      "post": {
        "summary": "Register a new user",
        "tags": ["Authentication"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "ninNumber": { "type": "string" },
                  "firstName": { "type": "string" },
                  "lastName": { "type": "string" },
                  "email": { "type": "string" },
                  "password": { "type": "string" }
                },
                "required": ["ninNumber", "firstName", "lastName", "email", "password"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User registered successfully"
          },
          "400": {
            "description": "Bad request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/login": {
      "post": {
        "summary": "User login",
        "tags": ["Authentication"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "ninNumber": { "type": "string" },
                  "password": { "type": "string" }
                },
                "required": ["ninNumber", "password"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/nin/{nin}": {
      "get": {
        "summary": "Get user by NIN number",
        "tags": ["Users"],
        "parameters": [
          {
            "name": "nin",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "NIN number of the user"
          }
        ],
        "responses": {
          "200": {
            "description": "User found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "404": {
            "description": "User not found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/bvn/{bvn}": {
      "get": {
        "summary": "Get user by BVN number",
        "tags": ["Users"],
        "parameters": [
          {
            "name": "bvn",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "BVN number of the user"
          }
        ],
        "responses": {
          "200": {
            "description": "User found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "404": {
            "description": "User not found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/bvn": {
      "get": {
        "summary": "Get all BVN numbers",
        "tags": ["BVN"],
        "responses": {
          "200": {
            "description": "List of BVN numbers",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/nin": {
      "get": {
        "summary": "Get all NIN numbers",
        "tags": ["NIN"],
        "responses": {
          "200": {
            "description": "List of NIN numbers",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/users": {
      "get": {
        "summary": "Get all users",
        "tags": ["Users"],
        "responses": {
          "200": {
            "description": "List of users",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/User"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/users/{id}": {
      "get": {
        "summary": "Get user by ID",
        "tags": ["Users"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "User ID"
          }
        ],
        "responses": {
          "200": {
            "description": "User found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "404": {
            "description": "User not found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/seed-nin": {
      "post": {
        "summary": "Seed NIN data",
        "tags": ["Seed Data"],
        "responses": {
          "201": {
            "description": "NIN data seeded successfully"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/seed-bvn": {
      "post": {
        "summary": "Seed BVN data",
        "tags": ["Seed Data"],
        "responses": {
          "201": {
            "description": "BVN data seeded successfully"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/seed-users": {
      "post": {
        "summary": "Seed user data",
        "tags": ["Seed Data"],
        "responses": {
          "201": {
            "description": "User data seeded successfully"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "ninNumber": {
            "type": "string"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        },
        "required": ["id", "ninNumber", "firstName", "lastName", "email"]
      }
    }
  }
}